+++ src/inputmanager.cpp
@@ -35,8 +35,8 @@ input_combo[20] = {POWER}; // eegg
 }
 
 InputManager::~InputManager() {
-	for (uint x = 0; x < joysticks.size(); x++)
-		if(SDL_JoystickOpened(x))
+	for (uint32_t x = 0; x < joysticks.size(); x++)
+		if (SDL_JoystickOpened(x))
 			SDL_JoystickClose(joysticks[x]);
 }
 
@@ -178,8 +178,8 @@ bool InputManager::update(bool wait) {
 		events.push_back(evcopy);
 	}
 
-	long now = SDL_GetTicks();
-	for (uint x = 0; x < actions.size(); x++) {
+	int32_t now = SDL_GetTicks();
+	for (uint32_t x = 0; x < actions.size(); x++) {
 		actions[x].active = isActive(x);
 		if (actions[x].active) {
 			memcpy(input_combo, input_combo + 1, sizeof(input_combo) - 1); // eegg
@@ -204,7 +204,7 @@ bool InputManager::combo() { // eegg
 }
 
 void InputManager::dropEvents() {
-	for (uint x = 0; x < actions.size(); x++) {
+	for (uint32_t x = 0; x < actions.size(); x++) {
 		actions[x].active = false;
 		if (actions[x].timer != NULL) {
 			SDL_RemoveTimer(actions[x].timer);
@@ -213,7 +213,7 @@ void InputManager::dropEvents() {
 	}
 }
 
-Uint32 InputManager::checkRepeat(Uint32 interval, void *_data) {
+uint32_t InputManager::checkRepeat(uint32_t interval, void *_data) {
 	RepeatEventData *data = (RepeatEventData *)_data;
 	InputManager *im = (class InputManager*)data->im;
 	SDL_JoystickUpdate();
@@ -261,9 +261,9 @@ int InputManager::count() {
 
 void InputManager::setInterval(int ms, int action) {
 	if (action < 0)
-		for (uint x = 0; x < actions.size(); x++)
+		for (uint32_t x = 0; x < actions.size(); x++)
 			actions[x].interval = ms;
-	else if ((uint)action < actions.size())
+	else if ((uint32_t)action < actions.size())
 		actions[action].interval = ms;
 }
 
@@ -275,7 +275,7 @@ void InputManager::setWakeUpInterval(int ms) {
 		wakeUpTimer = SDL_AddTimer(ms, wakeUp, NULL);
 }
 
-Uint32 InputManager::wakeUp(Uint32 interval, void *_data) {
+uint32_t InputManager::wakeUp(uint32_t interval, void *_data) {
 	SDL_Event *event = new SDL_Event();
 	event->type = SDL_WAKEUPEVENT;
 	SDL_PushEvent( event );
@@ -283,7 +283,7 @@ Uint32 InputManager::wakeUp(Uint32 interval, void *_data) {
 }
 
 bool &InputManager::operator[](int action) {
-	// if (action<0 || (uint)action>=actions.size()) return false;
+	// if (action<0 || (uint32_t)action>=actions.size()) return false;
 	return actions[action].active;
 }
 
@@ -305,7 +305,7 @@ bool InputManager::isActive(int action) {
 				}
 			break;
 			case InputManager::MAPPING_TYPE_KEYPRESS:
-				Uint8 *keystate = SDL_GetKeyState(NULL);
+				uint8_t *keystate = SDL_GetKeyState(NULL);
 				return keystate[map.value];
 			break;
 		}
